+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Fix: insert the Unreleased changelog header before the first version section
+
 ## 1.0.0 - 2026-09-17
 
 - feat: add initial action logic